+++ app/src/main/java/nextstep/signup/model/RegexValidation.kt
@@ -1,0 +1,11 @@
+package nextstep.signup.model
+
+class RegexValidation(
+    private val regex: Regex,
+    private val errorMessage: String,
+) : Validation {
+    override fun validate(text: String): Boolean =
+        regex.matches(text)
+
+    override fun errorMessage(text: String): String = errorMessage
+}
