+++ app/src/main/java/nextstep/signup/model/Validation.kt
@@ -1,0 +1,6 @@
+package nextstep.signup.model
+
+interface Validation {
+    fun validate(text: String): Boolean
+    fun errorMessage(text: String): String
+}
